refactor hw3
