--- hw3/app1/log_mock_test.txt
@@ -1,14 +1,0 @@
-time="2021-04-14 10:53:03" level=info msg="Reading dir: Dir1"
-time="2021-04-14 10:53:03" level=info msg="Reading dir: Dir2"
-time="2021-04-14 10:53:03" level=info msg="Reading dir: Dir1"
-time="2021-04-14 10:53:03" level=info msg="Reading dir: Dir2"
-time="2021-04-14 10:53:03" level=info msg="Reading dir: Dir1"
-time="2021-04-14 10:53:03" level=info msg="Reading dir: Dir1"
-time="2021-04-14 10:53:03" level=info msg="Problem with reading file: //Dir1/file2.txt stat //Dir1/file2.txt: no such file or directory"
-time="2021-04-14 10:53:03" level=info msg="Problem with reading file: //Dir2/Dir2/file2.txt stat //Dir2/Dir2/file2.txt: no such file or directory"
-time="2021-04-14 10:53:03" level=info msg="Problem with reading file: //Dir1/Dir1/file1.txt stat //Dir1/Dir1/file1.txt: no such file or directory"
-time="2021-04-14 10:53:03" level=info msg="Problem with reading file: //Dir1/Dir1/file2.txt stat //Dir1/Dir1/file2.txt: no such file or directory"
-time="2021-04-14 10:53:03" level=info msg="Problem with reading file: //Dir2/Dir1/file1.txt stat //Dir2/Dir1/file1.txt: no such file or directory"
-time="2021-04-14 10:53:03" level=info msg="Problem with reading file: //Dir2/Dir1/file2.txt stat //Dir2/Dir1/file2.txt: no such file or directory"
-time="2021-04-14 10:53:03" level=info msg="Problem with reading file: //Dir2/Dir2/Dir1/file1.txt stat //Dir2/Dir2/Dir1/file1.txt: no such file or directory"
-time="2021-04-14 10:53:03" level=info msg="Problem with reading file: //Dir2/Dir2/Dir1/file2.txt stat //Dir2/Dir2/Dir1/file2.txt: no such file or directory"
